Print errors to stdout
Naemons plugin API states that one must print at least one line of
output to stdout. Previous to this commit, only output to stderr would
be printed on errors, which meant the plugin did not conform to the
plugin standard.

With this commit we print everything to stdout.

This fixes: MON-12456

Signed-off-by: Jacob Hansen <jhansen@op5.com>