Add command line paramerer to exclude specified resource names from the check result
This commit adds the --ignore command line parameter to enable the users
to ignore or exclude specified resource names from being included in the
check result. The ignored resoures will not affect the final status being
reported by the plugin. The --ignore takes a regular expression and uses
python's re module to evaluate which resources will be ignored.

This resolves MON-13152.

Signed-off-by: Jerson Dumalaon <jdumalaon@itrsgroup.com>